chore: factor out accordion rendering
