+++ packages/accordion/src/Accordion.ts
@@ -13,103 +13,19 @@ governing permissions and limitations under the License.
 import {
     CSSResultArray,
     html,
-    PropertyValues,
-    SizedMixin,
-    SpectrumElement,
     TemplateResult,
 } from '@spectrum-web-components/base';
-import {
-    property,
-    queryAssignedNodes,
-} from '@spectrum-web-components/base/src/decorators.js';
-import { FocusGroupController } from '@spectrum-web-components/reactive-controllers/src/FocusGroup.js';
 
-import { AccordionItem } from './AccordionItem.js';
-
+import { AccordionBase } from './AccordionBase.js';
 import styles from './accordion.css.js';
 
 /**
  * @element sp-accordion
  * @slot - The sp-accordion-item children to display.
  */
-export class Accordion extends SizedMixin(SpectrumElement, {
-    noDefaultSize: true,
-}) {
+export class Accordion extends AccordionBase {
     public static override get styles(): CSSResultArray {
         return [styles];
-    }
-
-    /**
-     * Allows multiple accordion items to be opened at the same time
-     */
-    @property({ type: Boolean, reflect: true, attribute: 'allow-multiple' })
-    public allowMultiple = false;
-
-    /**
-     * Sets the spacing between the content to borders of an accordion item
-     */
-    @property({ type: String, reflect: true })
-    public density?: 'compact' | 'spacious';
-
-    @queryAssignedNodes()
-    private defaultNodes!: NodeListOf<AccordionItem>;
-
-    private get items(): AccordionItem[] {
-        return [...(this.defaultNodes || [])].filter(
-            (node: HTMLElement) => typeof node.tagName !== 'undefined'
-        ) as AccordionItem[];
-    }
-
-    focusGroupController = new FocusGroupController<AccordionItem>(this, {
-        direction: 'vertical',
-        elements: () => this.items,
-        isFocusableElement: (el: AccordionItem) => !el.disabled,
-    });
-
-    public override focus(): void {
-        this.focusGroupController.focus();
-    }
-
-    private async onToggle(event: Event): Promise<void> {
-        const target = event.target as AccordionItem;
-        // Let the event pass through the DOM so that it can be
-        // prevented from the outside if a user so desires.
-        await 0;
-        if (this.allowMultiple || event.defaultPrevented) {
-            // No toggling when `allowMultiple` or the user prevents it.
-            return;
-        }
-        const items = [...this.items] as AccordionItem[];
-        /* c8 ignore next 3 */
-        if (items && !items.length) {
-            // no toggling when there aren't items.
-            return;
-        }
-        items.forEach((item) => {
-            if (item !== target) {
-                // Close all the items that didn't dispatch the event.
-                item.open = false;
-            }
-        });
-    }
-
-    private handleSlotchange(): void {
-        this.focusGroupController.clearElementCache();
-        this.items.forEach((item) => {
-            item.size = this.size;
-        });
-    }
-
-    protected override updated(changed: PropertyValues<this>): void {
-        super.updated(changed);
-        if (
-            changed.has('size') &&
-            (!!changed.get('size') || this.size !== 'm')
-        ) {
-            this.items.forEach((item) => {
-                item.size = this.size;
-            });
-        }
     }
 
     protected override render(): TemplateResult {
