+++ packages/accordion/src/AccordionItem.ts
@@ -13,16 +13,13 @@ governing permissions and limitations under the License.
 import {
     CSSResultArray,
     html,
-    PropertyValues,
-    SizedMixin,
     TemplateResult,
 } from '@spectrum-web-components/base';
-import { property } from '@spectrum-web-components/base/src/decorators.js';
-import { Focusable } from '@spectrum-web-components/shared/src/focusable.js';
 import { when } from '@spectrum-web-components/base/src/directives.js';
 import '@spectrum-web-components/icons-ui/icons/sp-icon-chevron100.js';
 import chevronIconStyles from '@spectrum-web-components/icon/src/spectrum-icon-chevron.css.js';
 
+import { AccordionItemBase } from './AccordionItemBase.js';
 import styles from './accordion-item.css.js';
 
 const chevronIcon: Record<string, () => TemplateResult> = {
@@ -65,46 +62,9 @@ const chevronIcon: Record<string, () => TemplateResult> = {
  * @slot - The content of the item that is hidden when the item is not open
  * @fires sp-accordion-item-toggle - Announce that an accordion item has been toggled while allowing the event to be cancelled.
  */
-export class AccordionItem extends SizedMixin(Focusable, {
-    noDefaultSize: true,
-}) {
+export class AccordionItem extends AccordionItemBase {
     public static override get styles(): CSSResultArray {
         return [styles, chevronIconStyles];
-    }
-
-    @property({ type: Boolean, reflect: true })
-    public open = false;
-
-    @property({ type: String, reflect: true })
-    public label = '';
-
-    @property({ type: Boolean, reflect: true })
-    public override disabled = false;
-
-    public override get focusElement(): HTMLElement {
-        return this.shadowRoot.querySelector('#header') as HTMLElement;
-    }
-
-    private onClick(): void {
-        /* c8 ignore next 3 */
-        if (this.disabled) {
-            return;
-        }
-        this.toggle();
-    }
-
-    private toggle(): void {
-        this.open = !this.open;
-        const applyDefault = this.dispatchEvent(
-            new CustomEvent('sp-accordion-item-toggle', {
-                bubbles: true,
-                composed: true,
-                cancelable: true,
-            })
-        );
-        if (!applyDefault) {
-            this.open = !this.open;
-        }
     }
 
     protected renderChevronIcon = (): TemplateResult => {
@@ -129,16 +89,5 @@ export class AccordionItem extends SizedMixin(Focusable, {
                 <slot></slot>
             </div>
         `;
-    }
-
-    protected override updated(changes: PropertyValues): void {
-        super.updated(changes);
-        if (changes.has('disabled')) {
-            if (this.disabled) {
-                this.setAttribute('aria-disabled', 'true');
-            } else {
-                this.removeAttribute('aria-disabled');
-            }
-        }
     }
 }
